+++ src/valency_anndata/datasets/__init__.py
@@ -5,6 +5,7 @@ from ._load_bg2050 import bg2050
 from ._load_chile_protest import chile_protest
 from ._load_cuba_protest import cuba_protest
 from ._load_klimarat import klimarat
+from ._load_vtaiwan import vtaiwan
 
 __all__ = [
     "load",
@@ -14,5 +15,6 @@ __all__ = [
     "chile_protest",
     "cuba_protest",
     "klimarat",
+    "vtaiwan",
     "translate_statements",
 ]